requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
aiofiles==23.2.0
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.0
fake-useragent==1.4.0
tenacity==8.2.3
selenium==4.15.0
loguru==0.7.2
openai>=1.0.0
